- Provide HttpClient for native; - Onboard test activities: comment direction session dependencies(because one was moved to code module).
