+++ app/src/main/java/com/mapbox/services/android/navigation/testapp/activity/OnboardRouterActivityKt.kt
@@ -1,5 +1,7 @@
 package com.mapbox.services.android.navigation.testapp.activity
 
+// import com.mapbox.navigation.base.route.DirectionsSession
+// import com.mapbox.navigation.directions.session.MapboxDirectionsSession
 import android.os.Bundle
 import android.os.Environment
 import android.widget.Toast
@@ -13,35 +15,30 @@ import com.mapbox.mapboxsdk.geometry.LatLng
 import com.mapbox.mapboxsdk.maps.MapboxMap
 import com.mapbox.mapboxsdk.maps.OnMapReadyCallback
 import com.mapbox.mapboxsdk.maps.Style
-import com.mapbox.navigation.base.route.DirectionsSession
 import com.mapbox.navigation.base.route.Router
-import com.mapbox.navigation.base.route.model.Route
 import com.mapbox.navigation.base.route.model.RouteOptionsNavigation
-import com.mapbox.navigation.directions.session.MapboxDirectionsSession
 import com.mapbox.navigation.route.onboard.MapboxOnboardRouter
 import com.mapbox.navigation.route.onboard.model.Config
 import com.mapbox.navigation.utils.extensions.ifNonNull
 import com.mapbox.services.android.navigation.testapp.R
 import com.mapbox.services.android.navigation.testapp.utils.Utils
 import com.mapbox.services.android.navigation.ui.v5.route.NavigationMapRoute
-import com.mapbox.services.android.navigation.v5.utils.extensions.mapToDirectionsRoute
 import com.mapbox.turf.TurfConstants
 import com.mapbox.turf.TurfMeasurement
 import java.io.File
 import kotlinx.android.synthetic.main.activity_mock_navigation.*
-import timber.log.Timber
 
 class OnboardRouterActivityKt : AppCompatActivity(),
     OnMapReadyCallback,
-    MapboxMap.OnMapClickListener,
-    DirectionsSession.RouteObserver {
+    MapboxMap.OnMapClickListener
+// DirectionsSession.RouteObserver {
 
     private lateinit var onboardRouter: Router
     private lateinit var mapboxMap: MapboxMap
 
     private var route: DirectionsRoute? = null
     private lateinit var navigationMapRoute: NavigationMapRoute
-    private lateinit var directionsSession: DirectionsSession
+    // private lateinit var directionsSession: DirectionsSession
     private var origin: Point? = null
     private var destination: Point? = null
     private var waypoint: Point? = null
@@ -62,12 +59,12 @@ class OnboardRouterActivityKt : AppCompatActivity(),
         )
         val fileTiles = File(file, "tiles")
         val config = Config(
-                fileTiles.absolutePath,
-                null,
-                null,
-                null,
-                null // working with pre-fetched tiles only
-            )
+            fileTiles.absolutePath,
+            null,
+            null,
+            null,
+            null // working with pre-fetched tiles only
+        )
         onboardRouter = MapboxOnboardRouter(config)
     }
 
@@ -104,10 +101,10 @@ class OnboardRouterActivityKt : AppCompatActivity(),
     }
 
     private fun findRoute() {
-        directionsSession = MapboxDirectionsSession(
-            onboardRouter,
-            this
-        )
+        // directionsSession = MapboxDirectionsSession(
+        //     onboardRouter,
+        //     this
+        // )
         ifNonNull(origin, destination) { origin, destination ->
             if (TurfMeasurement.distance(origin, destination, TurfConstants.UNIT_METERS) > 50) {
 
@@ -118,7 +115,7 @@ class OnboardRouterActivityKt : AppCompatActivity(),
                         .destination(destination)
                 waypoint?.let { optionsBuilder.addWaypoint(it) }
 
-                directionsSession.requestRoutes(optionsBuilder.build())
+                // directionsSession.requestRoutes(optionsBuilder.build())
             }
         }
     }
@@ -136,7 +133,11 @@ class OnboardRouterActivityKt : AppCompatActivity(),
                 findRoute()
             }
             else -> {
-                Toast.makeText(this, "Only 2 waypoints supported for this example", Toast.LENGTH_LONG)
+                Toast.makeText(
+                    this,
+                    "Only 2 waypoints supported for this example",
+                    Toast.LENGTH_LONG
+                )
                     .show()
                 clearMap()
             }
@@ -147,20 +148,20 @@ class OnboardRouterActivityKt : AppCompatActivity(),
     /*
      * DirectionSessions.RouteObserver
      */
-    override fun onRoutesChanged(routes: List<Route>) {
-        if (routes.isNotEmpty()) {
-            route = routes[0].mapToDirectionsRoute()
-            navigationMapRoute.addRoute(route)
-        }
-    }
+    // override fun onRoutesChanged(routes: List<Route>) {
+    //     if (routes.isNotEmpty()) {
+    //         route = routes[0].mapToDirectionsRoute()
+    //         navigationMapRoute.addRoute(route)
+    //     }
+    // }
 
-    override fun onRoutesRequested() {
-        Timber.d("onRoutesRequested: navigation.getRoute()")
-    }
-
-    override fun onRoutesRequestFailure(throwable: Throwable) {
-        Timber.e(throwable, "onRoutesRequestFailure: navigation.getRoute()")
-    }
+    // override fun onRoutesRequested() {
+    //     Timber.d("onRoutesRequested: navigation.getRoute()")
+    // }
+    //
+    // override fun onRoutesRequestFailure(throwable: Throwable) {
+    //     Timber.e(throwable, "onRoutesRequestFailure: navigation.getRoute()")
+    // }
 
     /*
      * Activity lifecycle methods
@@ -192,7 +193,7 @@ class OnboardRouterActivityKt : AppCompatActivity(),
 
     override fun onDestroy() {
         super.onDestroy()
-        directionsSession.cancel()
+        // directionsSession.cancel()
         mapboxMap.removeOnMapClickListener(this)
         mapView.onDestroy()
     }
